backend/api/requirements.txt
fastapi
uvicorn[standard]
python-multipart
sqlmodel
python-dotenv
bcrypt
pyjwt
faiss-cpu
transformers
sentence-transformers
aiofiles
requests
PyPDF2
